requirements.txt
Django
django-cors-headers
djangorestframework
PyJWT
